+++ L3gion/src/L3gion/Core/Timestep.h
@@ -5,16 +5,16 @@ namespace L3gion
 	class Timestep
 	{
 	public:
-		Timestep(float time = 0.0f)
+		Timestep(double time = 0.0f)
 			: m_Time(time)
 		{
 		}
 
-		operator float() const { return m_Time; }
+		operator double() const { return m_Time; }
 
-		float getSeconds() const { return m_Time; }
-		float getMiliSeconds() const { return m_Time * 1000.0f; }
+		double getSeconds() const { return m_Time; }
+		double getMiliSeconds() const { return m_Time * 1000.0f; }
 	private:
-		float m_Time;
+		double m_Time;
 	};
 }
